requirements.txt
certifi==2020.11.8
chardet==3.0.4
future==0.18.2
idna==2.10
oauthlib==3.1.0
peewee==3.14.0
PySocks==1.7.1
python-telegram-bot==5.0.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.5
python-dotenv